+++ extensions/azure/cosmos/policy-store-cosmos/src/main/java/org/eclipse/dataspaceconnector/cosmos/policy/store/CosmosPolicyStoreExtension.java
@@ -12,41 +12,41 @@
  *
  */
 
-package org.eclipse.dataspaceconnector.contract.definition.store;
+package org.eclipse.dataspaceconnector.cosmos.policy.store;
 
 import net.jodah.failsafe.RetryPolicy;
 import org.eclipse.dataspaceconnector.azure.cosmos.CosmosDbApiImpl;
-import org.eclipse.dataspaceconnector.cosmos.policy.store.model.ContractDefinitionDocument;
-import org.eclipse.dataspaceconnector.dataloading.ContractDefinitionLoader;
-import org.eclipse.dataspaceconnector.spi.contract.offer.store.ContractDefinitionStore;
+import org.eclipse.dataspaceconnector.spi.policy.store.PolicyStore;
 import org.eclipse.dataspaceconnector.spi.security.Vault;
+import org.eclipse.dataspaceconnector.spi.system.Inject;
 import org.eclipse.dataspaceconnector.spi.system.Provides;
 import org.eclipse.dataspaceconnector.spi.system.ServiceExtension;
 import org.eclipse.dataspaceconnector.spi.system.ServiceExtensionContext;
 import org.eclipse.dataspaceconnector.spi.system.health.HealthCheckService;
 
-@Provides({ ContractDefinitionStore.class, ContractDefinitionLoader.class })
-public class CosmosContractDefinitionStoreExtension implements ServiceExtension {
+@Provides({ PolicyStore.class })
+public class CosmosPolicyStoreExtension implements ServiceExtension {
+
+    @Inject
+    private RetryPolicy retryPolicy;
+    @Inject
+    private Vault vault;
 
     @Override
     public String name() {
-        return "CosmosDB ContractDefinition Store";
+        return "CosmosDB Policy Store";
     }
 
     @Override
     public void initialize(ServiceExtensionContext context) {
-        var configuration = new CosmosContractDefinitionStoreConfig(context);
-        Vault vault = context.getService(Vault.class);
+        var configuration = new CosmosPolicyStoreConfig(context);
 
         var cosmosDbApi = new CosmosDbApiImpl(vault, configuration);
 
-        var store = new CosmosContractDefinitionStore(cosmosDbApi, context.getTypeManager(), (RetryPolicy<Object>) context.getService(RetryPolicy.class), configuration.getPartitionKey());
-        context.registerService(ContractDefinitionStore.class, store);
+        var store = new CosmosPolicyStore(cosmosDbApi, context.getTypeManager(), retryPolicy, configuration.getPartitionKey());
+        context.registerService(PolicyStore.class, store);
 
-        ContractDefinitionLoader loader = store::save;
-        context.registerService(ContractDefinitionLoader.class, loader);
-
-        context.getTypeManager().registerTypes(ContractDefinitionDocument.class);
+        context.getTypeManager().registerTypes(PolicyDocument.class);
 
         context.getService(HealthCheckService.class).addReadinessProvider(() -> cosmosDbApi.get().forComponent(name()));