+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+
 # Changelog
 
 All notable changes to this project will be documented in this file - formatted and maintained according to the rules
@@ -43,6 +44,7 @@ the detailed section referring to by linking pull requests or issues.
 * In-memory implementation of PolicyStore (#930)
 * Implement AssetLoader, AssetIndex, DataAddressResolver for SQL (#863)
 * Let Control Plane delegate data transfer to Data Plane (#988)
+* CosmosDb based `PolicyStore` (#826)
 
 #### Changed